add circular dep check, fix circulars
