+++ packages/core/src/context/types.ts
@@ -1,7 +1,7 @@
-import { PlatformKey, RenativeConfigFile, RenativeConfigLocal } from '../schema/ts/types';
-import { RnvEngine, RnvEnginePlatform } from '../engines/types';
-import { OverridesOptions } from '../system/types';
-import { RnvPlatform } from '../types';
+import type { PlatformKey, RenativeConfigFile, RenativeConfigLocal } from '../schema/ts/types';
+import type { RnvEngine, RnvEnginePlatform } from '../engines/types';
+import type { OverridesOptions } from '../system/types';
+import type { RnvPlatform } from '../types';
 
 export interface RnvContext<Payload = any> {
     /**
